Metadata sending and searching be done against the write server.
